moduloClientes: sweetAlert y algunos estilos

+++ src/app/admin/cliente/cliente-tarifa/cliente-tarifa.component.ts
@@ -7,6 +7,7 @@ import { AdicionalKm, AdicionalTarifa, CargasGenerales, TarifaCliente, TarifaEsp
 import { DbFirestoreService } from 'src/app/servicios/database/db-firestore.service';
 import { StorageService } from 'src/app/servicios/storage/storage.service';
 import { ModalAltaTarifaComponent } from '../modal-alta-tarifa/modal-alta-tarifa.component';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-cliente-tarifa',
@@ -124,7 +125,26 @@ this.acompanianteEditForm = this.fb.group({
   }
 
   eliminarTarifa(tarifa:TarifaCliente){    
-    this.storageService.deleteItem(this.componente, tarifa);
+    Swal.fire({
+      title: "¿Eliminar la tarifa?",
+      text: "No se podrá revertir esta acción",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Confirmar",
+      cancelButtonText: "Cancelar"
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.storageService.deleteItem(this.componente, tarifa);
+        Swal.fire({
+          title: "Confirmado",
+          text: "La tarifa ha sido borrada",
+          icon: "success"
+        });
+      }
+    });   
+    
   }
 
   armarTarifaEditar(){
@@ -154,7 +174,26 @@ this.acompanianteEditForm = this.fb.group({
   }
 
   onSubmitEdit(){
-    this.armarTarifaModificada();
+    Swal.fire({
+      title: "¿Guardar los cambios?",
+      text: "No se podrá revertir esta acción",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Confirmar",
+      cancelButtonText: "Cancelar"
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.armarTarifaModificada();
+        Swal.fire({
+          title: "Confirmado",
+          text: "Los cambios se han guardado",
+          icon: "success"
+        });
+      }
+    }); 
+    
   }
 
   armarTarifaModificada(){
